pyproject.toml
[project]
name = "llmcompanioncord"
version = "0.1.0"
description = "An LLM companion for your Discord!"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.4",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llmcompanioncord"]

[project.scripts]
llmcompanioncord = "llmcompanioncord.main:main"
